(RSP) MIPS offsets are always sign-extended in memory loads/stores.
Part of the purpose of these commits is to get non-portable types (like DWORD) to fixed-size, more readable types that conform to MIPS size types.  While the usage of Windows API types was a part of the problem, it also is not portable to repeatedly say (short)(union.offset) every single time throughout the source, when the MIPS instruction set architecture automatically sign-extends the 16-bit offset value anyway regardless when doing any of these memory loads and stores.  As such, it was counter-intuitive to repeatedly use type conversions to the (short) type, when (short) is not portable either and not necessarily a 16-bit type.  Now that the `offset` member of the RSP opcode union is already re-defined a int16_t (actually a "signed" 16-bit struct bitfield, which is even better), it seems more intuitive to just go with that universally for MIPS offset decoding.  "Find All References" in VS2013 did not detect any instances of referring to `offset` without an explicit type-cast to (short) with only some inconsistent exceptions near the end of the RSP commands debugger that sprintf'd it as a %04X.

I did not make this change outside of the RSP plugin to the main Project64 R4300 CPU source, just the RSP for now.  In a future commit, this change should probably be applied to the main Project64 CPU as well.