Release date update for SW - Shadows of the Empire
http://en.wikipedia.org/wiki/Star_Wars:_Shadows_of_the_Empire_%28video_game%29
I got the correct release dates for (E) and (J).

http://www.gamefaqs.com/n64/198789-star-wars-shadows-of-the-empire/data

12/03/1996  NUS-NSWE-USA      Star Wars: Shadows of the Empire
Date matches v1.0!

01/26/1999  NUS-NSWE-USA-1    Star Wars: Shadows of the Empire (Players Choice)
so is this v1.1 or v1.2,
since this is the only other version for North America... v1.2 as probably

I redid dates as follows
(E) = 1997/03/01
(J) = 1997/06/14
v1.1 = 199x
v1.2 = 01/26/1999  = as dated in rdx 1999/01/26

I found a reference  "NUS-006 (USA) | NUS-NSWE-USA-1" to v1.2
on a site so its definitely v1.2