Load null bytes as whitespace.
A few lines down in this file, there is a while() loop that removes all trailing spaces to replace them with null bytes, so "SUPER MARIO 64      " becomes "SUPER MARIO 64", whether or not we converted null bytes to spaces.  Also, the null byte is not documented to be a valid character for game developers to use in their ROM headers, even though for a few ROMs this is being done anyway.