Slightly less broken Indiana Jones VI\AI
The US and PAL versions behave quite differently when it comes to refusing to boot randomly. The ideal VI should be around 1500, but the closest I can get seems around 1800 for PAL and 2050 for US. US version runs faster - significantly so in certain levels. However, this could be related to the PAL conversion - although I doubt it.