[Common] some scope and spacing formatting clean-ups
Miserably hated doing this commit.  Couldn't tell which code was whose, which was copyrighted, which was foreign enough to Project64 that I'd probably best just leave-as is, which was even worth considering part of Project64, which cleanups to omit doing and ignore because some ugly practices were rampant throughout the entire file and distracted from the purpose of this pull request too much.  So tried to stick to mostly just the braces/indentation changes here.