Do not capitalize "on".
"On" is a preposition shorter than 5 letters long.  Standard title case does not ever capitalize short prepositions (unless they are the very first word of the title, then usually).  You can see this logic in the Visual Studio 2008 IDE itself:  It has menu items such as "Attach to Process" and "Find[/Replace] in Files", in which the prepositions "to" and "in" are, correctly so, not capitalized by Microsoft in the VS2008 menus.