drop '%'
Either say "Percentage", or just drop the '%' since this symbol of unit is not relevant to what the menu item should convey to the user.  What we're really trying to do is show the CPU usage statistics...most likely the users would see for themselves what unit it comes out as (percentage, fraction, whatever).