used in a complete sentence:  "Sally gave me a short cut."
Did she give me a detour kind of "short cut", or did she just cut off my balls?
If we prefer not to say "shortcut", then at least say "short-cut".  This way, the adjective "short" is co-joined with "cut" to indicate that they are connected.  This guarantees that the "cut" is associated with "short", and not whatever word comes after.