Alphabetical order fix
Alphabetical Order

The English alphabet starts with the letter a and finishes with the
letter z. We always write the alphabet in the same order:

a-b-c-d-e-f-g-h-i-j-k-l-m-n-o-p-q-r-s-t-u-v-w-x-y-z

A-B-C-D-E-F-G-H-I-J-K-L-M-N-O-P-Q-R-S-T-U-V-W-X-Y-Z

This order is called "alphabetical order". We often write lists in
alphabetical order. For example, to write a list of countries in
alphabetical order, we start with countries that begin with the letter
A, then with countries that begin with B, and so on. For example:

Australia
Brazil
Canada