[Plugin Specs] reverted contr 1.1 for backward-compatibility
No plugins for Windows (that can run on Project64) currently use the 1.1 controller plugin specifications, and the ones that do use it (for Linux) use the publicly released spec which did not have the pointer version in the late-modified spec.

We don't want to let plugin wars roam between the emulators' implementations.