Allow for pushing two C Buttons at once when using an analog stick
This value specifies how far an analog stick has to be pressed to count as a C Button pres when c Buttons are mapped to an analog stick.  The old value did not allow for pressing two c buttons at once, because diagonals did not meet the value.  This value specifies that an analog stick must be pressed roughly halfway to count as a C Button press, which allows you to push two C Buttons at once.