[Common] leftover remnant of Win32 command shell
I found out the hard way while getting `libglitch64.a` to link into PJ64 Glide64 on Linux that I was getting an 8-byte static library file.  This was because the Unix shell variable `$OBJ_LIST` was null, due to using the `set` keyword which was a result of lazy copy pasta while porting my MinGW script to compile the PJ64 Common library on Windows.