[RDB] Use best possible settings for Super Mario 64
Starting work on the RDB. First things first: Super Mario 64. This is like the most popular game for the console, and probably what most people will first play on the emulator, so it should run as good as possible. Explanation of these settings:

1. Set Counter Factor to 1. Makes the title screen with Mario's disembodied head far smoother (runs at a steady 30 DL/s NTSC/25 DL/s PAL), and should prevent any slowdowns. Also makes transitions between maps slightly quicker.
2. Disabled all Self-Modifying Code Methods. 99% certain this game doesn't need them. Should increase core speed, not that this game is resource-intensive but it doesn't harm to optimize.

Tested these settings very thoroughly with the (U) ROM (collected all 120 stars, did about every action possible, even tried the hidden level select and debug displays), and fairly extensively with the (E), (J) and Shindou Edition ROMs.

I was tempted to disable Advanced Block Linking (also tested fairly extensively with it Off) to prevent microstuttering, but I think that's something that should be considered on a more global level.