[README.md] Reword and remove superfluous wording
I liked some of the changes by @Alen-Alic but I think keeping it short and sweet is best
Under the requisites including sound card and storage is redundant, since every computer that runs Windows XP comes with audio capabilities, and storage is not a concern since who doesn't have 5MB of storage available.

After this commit, please do not add other modifications to the README.md unless there has been a significant development (like Android version coming back). For the meantime, we believe it is already good as-is.