Remove greedy Good Name
Removes the Good Name from the greedy THE LEGEND OF ZELDA rdb entry.
Settings still seem to be retained between all relevant ROMs and
changing an option in one will change the option for all of them.

This is preferable as the file name is the only way to distinguish
randomizer ROMs in the first place. It also won't override unrelated
rom hacks with the wrong name. This should reduce user confusion for
both of these cases.