Fix WelcomeScreen setting Default_GFX_Plugin
Settings.cpp - Was missing a value for the new SettingID while running a
Debug build causing breakpoint to be handled when clicking OK. A new
handler with a default value has been added.

WelcomeScreen.cpp - Would set the Default_GFX_Plugin value and the value
would be used when launching a game or viewing the Plugins menu in
Options and set the Current_GFX_Plugin, but if a user were to try to
configure their Graphics Plugin prior to launching a game, they would
always open the Project64 Video config screen, even if they had chosen
GLideN64 as the default plugin. The Current_GFX_Plugin is now set with
the value of the Default_GFX_Plugin to ensure the correct configuration
dialog is opened.

WelcomeScreen.cpp - String value would always load the non-debug version
of Project64 Video after the above changes. The string to use is now
loaded from the Default_GFX_Plugin so if the default has never been set
the proper default value should now be used. If it has been set before
and is not a Project64 Video plugin, the value will be set to Project64
Video (non-debug) to keep the dialog accurate.