N64Rom: Load first found ROM in 7z file
Should now search the entire 7zip file looking for a valid N64 ROM file
and load the first one it finds. Left comments referring to a future
dialog for selecting which file you want to open as it seems like a
good idea, but I don't have the familiarity with WTL yet to try to add
it.

May be necessary to skip files of a certain size while opening this way
as Project64 will try to allocate memory for the "ROM" file before it
is able to check if it really is a ROM file.