General GitHub-related project improvements and updates
Added a CONTRIBUTING file to give new contributors an idea of what they can do to help with the project.

Also added an issue and pull request template to assist people with providing accurate and detailed bug reports, as well as making pull requests a bit easier to make and also read.

I also added a config.yml file to add a "sponsor" button that points to the Project64 "support the project" page on the website for people who want to support the project.

I added an EditorConfig file so that no matter what text editor or IDE contributors use,  we still have consistency across all contributions. I also added some EOL and charset enforcements as well as the final newline option to enforce our gitattributes file.

Speaking of the gitattributes file, I updated it with some new file types that were in the repo. I also added the generic auto normalize (it's the first line). This MAY cause the entire repo to look like it's being changed, but it isn't I promise. If this happens it will be the first and last time unless we change it again.

Most of the additions are just (as above) enforcing things we are already doing, like how diffs are handled and such.

I also added some future proofing with Linux and Mac-related file types and some other binary files.

This PR is still a work-in-progress and is mostly done very late and early into the morning, so be gentle with me. 😄