Enforce a standard issue template
This will enforce a standard issue template, and will force the user to make sure they are creating the issue with all of the info the developers and contributors need to help the user.

Also updates the README to be less confusing and also a bit more broad in some cases.